# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateBook:
    properties:
      author:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
  models.LoginUser:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.PatchBook:
    properties:
      author:
        type: string
      title:
        type: string
    type: object
  models.ReplaceBook:
    properties:
      author:
        type: string
      title:
        type: string
    required:
    - author
    - title
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 'Go/Gin REST API template: books CRUD, register/login, Redis-backed
    list cache, Postgres via GORM, Mongo access logs, rate limiting, and Swagger.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: golang-rest-api-template
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /books:
    get:
      description: Get a list of all books with optional pagination. List entries
        are keyed by a monotonic cache generation (no Redis KEYS) and canonical integer
        offset/limit after parsing (leading zeros and surrounding whitespace in query
        params do not fragment the cache). Concurrent cache misses for the same offset/limit
        and generation are coalesced (singleflight) so only one database read and
        Redis write runs per cache key.
      parameters:
      - default: 0
        description: Offset for pagination (must be >= 0)
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination (minimum 1, capped at 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of books
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all books with pagination
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a new book with the given input data
      parameters:
      - description: Create book object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateBook'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Create a new book
      tags:
      - books
  /books/{id}:
    delete:
      description: Delete the book with the given ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted book
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Delete a book by ID
      tags:
      - books
    get:
      description: Get details of a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved book
          schema:
            $ref: '#/definitions/models.Book'
        "404":
          description: Book not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Find a book by ID
      tags:
      - books
    patch:
      consumes:
      - application/json
      description: Updates only fields present in the JSON body (at least one of title
        or author).
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.PatchBook'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Partially update a book by ID (PATCH)
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Replaces both title and author. Use PATCH for partial updates.
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Full book fields
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ReplaceBook'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated book
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: book not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - JwtAuth: []
      summary: Replace a book by ID (PUT)
      tags:
      - books
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using username and password, returns a JWT
        token if successful
      parameters:
      - description: User login object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: JWT Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Authenticate a user
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the given username and password
      parameters:
      - description: User registration object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Register a new user
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"